refactor: move MST et al into module
Adds a "mst" (merkle search tree) submodule in anti_entropy, and moves
all the MST code into it.

This makes space for a gossip-based sync primitive to live here too.
